Proposed Fix for mBoardShip
Been playing around with the SRV macros and think I have found a solution for the mBoardShip macro. Primarily, this adds a couple more UIDown commands to ensure if its the first run (where it loops menu) it will hit the right spot, but also capture the right spot on subsequent runs.  Some minor timing changes too, but hopefully works as well for you as it does for me :)